requirements.txt
falcon==2.0.0
falcon-swagger-ui==1.1.2
gunicorn==19.9.0
google==2.0.2
google-oauth==1.0.1
google-auth==1.6.3
mysql-connector==2.2.9
yoyo-migrations==6.1.0
PyMySQL==0.9.3
